requirements.txt
python-telegram-bot==13.15
python-dotenv==0.21.0
pandas==1.5.2
openpyxl==3.0.10
